Jammy support (#40)
Add Ubuntu 22.04 Jammy support
* new install way

* add python3-apt

* add branch
* address missed instructions

+++ tools/ansible.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
 COOKBOOK_PATH=/tmp/codio_playbook
-BRANCH=master
+
+if test -z "${BRANCH}"; then
+    BRANCH=master
+fi
+
 
 CODENAME=$(lsb_release -c -s)
 PY_FLAG=$1
@@ -9,6 +13,7 @@ PY_FLAG=$1
 IS_TRUSTY() { [ "${CODENAME}" == "trusty" ]; }
 IS_XENIAL() { [ "${CODENAME}" == "xenial" ]; }
 IS_BIONIC() { [ "${CODENAME}" == "bionic" ]; }
+IS_JAMMY() { [ "${CODENAME}" == "jammy" ]; }
 USE_PYTHON3() { [ "${PY_FLAG}" == "python3" ]; }
 
 do_cmd()
@@ -56,10 +61,15 @@ if IS_BIONIC; then
     ansible --version | grep -q '2.7.5' 2> /dev/null
     is_ansible_right=$?
 fi
+if IS_JAMMY; then
+    ansible --version | grep -q '2.10' 2> /dev/null
+    is_ansible_right=$?
+fi
+
 if [ $is_ansible_right -ne 0 ]; then
     do_cmd sudo apt-get update
     if USE_PYTHON3; then
-      do_cmd sudo apt-get -y install python3
+      do_cmd sudo apt-get -y install python3 python3-apt
     fi
     if IS_TRUSTY; then
         do_cmd sudo apt-get -y install wget python python-support python-yaml python-httplib2 python-setuptools python-markupsafe python-jinja2 python-paramiko sshpass
@@ -73,8 +83,13 @@ if [ $is_ansible_right -ne 0 ]; then
         do_cmd sudo apt-get -y install wget python python-crypto python-yaml python-httplib2 python-setuptools python-markupsafe python-jinja2 python-paramiko sshpass
         do_cmd sudo wget -O /tmp/ansible.deb https://raw.githubusercontent.com/codio/install_software/${BRANCH}/tools/ansible_2.7.5-1ppa_bionic_all.deb
     fi
-    do_cmd sudo dpkg -i /tmp/ansible.deb
-    do_cmd sudo rm /tmp/ansible.deb
+    if IS_JAMMY; then
+        do_cmd sudo apt update
+        do_cmd sudo apt install -y ansible
+    else
+        do_cmd sudo dpkg -i /tmp/ansible.deb
+        do_cmd sudo rm /tmp/ansible.deb
+    fi
 fi
 
 download_playbook
